api_gateway/requirements.txt
alembic==1.14.0
apturl==0.5.2
asgiref==3.5.0
attrs==21.2.0
bcrypt==3.2.0
blinker==1.9.0
Brlapi==0.8.3
cached-property==1.5.2
certifi==2020.6.20
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.1.8
colorama==0.4.4
command-not-found==0.3
common==0.1.2
config==0.5.1
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
Deprecated==1.2.15
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==3.2.12
dnspython==2.1.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
fasteners==0.14.1
Flask==3.1.0
Flask-JWT==0.3.2
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.10.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
future==0.18.2
gpg==1.16.0
greenlet==3.1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.5
jsonschema==3.2.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
limits==4.0.0
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
ordered-set==4.1.0
packaging==24.2
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
prettytable==2.5.0
protobuf==3.12.4
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
Pygments==2.19.0
PyGObject==3.42.1
PyJWT==2.10.1
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
redis==5.2.1
reportlab==3.6.8
requests==2.25.1
requests-toolbelt==0.9.1
rich==13.9.4
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
setuptools==69.2.0
six==1.16.0
SQLAlchemy==2.0.36
sqlparse==0.4.2
ssh-import-id==5.11
systemd-python==234
texttable==1.6.4
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
wcwidth==0.2.5
websocket-client==1.2.3
Werkzeug==3.1.3
wheel==0.43.0
wrapt==1.17.0
xdg==5
xkit==0.0.0
zipp==1.0.0
